requirements.txt
python-telegram-bot
movieposters
BeautifulSoup4
urlextract
requests
instagrapi
Pillow
python-dotenv